Make webpack configuration more similar to common practice, as described on https://stackoverflow.com/a/43476629/3711110
