+++ module/eCampWeb/vue/.babelrc
@@ -1,5 +1,5 @@
 {
     "presets": [
-        [ "env", { "modules": false } ]
+        [ "@babel/preset-env", { "modules": false } ]
     ]
 }
